requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0
scipy>=1.7.0
tqdm>=4.62.0
requests>=2.25.0
urllib3>=1.26.0
